Safaar-SafarB1/src/main/java/com/capgemini/Safaar/DTO/BusDetails.java
package com.capgemini.Safaar.DTO;

public class BusDetails {
	private String source;
	private String destination;
	private String departureDate;
	private String departureTime;
	private String busType;
	private String travelsName;
	private String duration;
	private String rating;
	private String fare;
	
	
	public BusDetails(String source, String destination, String departureDate, String departureTime, String busType,
			String travelsName, String duration, String rating, String fare) {
		
		this.source = source;
		this.destination = destination;
		this.departureDate = departureDate;
		this.departureTime = departureTime;
		this.busType = busType;
		this.travelsName = travelsName;
		this.duration = duration;
		this.rating=rating;
		this.setFare(fare);
	}
	public String getSource() {
		return source;
	}
	public void setSource(String source) {
		this.source = source;
	}
	public String getDestination() {
		return destination;
	}
	public void setDestination(String destination) {
		this.destination = destination;
	}
	public String getDepartureDate() {
		return departureDate;
	}
	public void setDepartureDate(String departureDate) {
		this.departureDate = departureDate;
	}
	public String getDepartureTime() {
		return departureTime;
	}
	public void setDepartureTime(String departureTime) {
		this.departureTime = departureTime;
	}
	public String getBusType() {
		return busType;
	}
	public void setBusType(String busType) {
		this.busType = busType;
	}

	public String getTravelsName() {
		return travelsName;
	}

	public void setTravelsName(String travelsName) {
		this.travelsName = travelsName;
	}

	public String getDuration() {
		return duration;
	}

	public void setDuration(String duration) {
		this.duration = duration;
	}

	public String getRating() {
		return rating;
	}

	public void setRating(String rating) {
		this.rating = rating;
	}
	public String getFare() {
		return fare;
	}
	public void setFare(String fare) {
		this.fare = fare;
	}
	
}
